[DIRTY] Adds H5 protocol to the built kernel.
The rtl8723bs wifi module is configured via the userspace, using a TTY
device created for the bluetooth module. This is how many (cheaper?)
bluetooth module works, and is not that unusual. The userspace
configuration utility will only be able to communicate with the
bluetooth device if the kernel has the right protocols built.

This configuration option makes the H5 protocol available. It is
"hidden" under that other bluetooth device name since that one
apparently uses that protocol too.

*sigh*