WIP: Make platform accurate to intent
We're dropping `pkgsCross` here, as anyway it probably did nothing since
`CROSS_CM3` used `gcc-arm-embedded`. Generally `pkgsCross` is not
needed within derivations.

We're marking the platform with the proper target. Though yes it seems
to build fine with a simple `nix-build -A mox-boot-builder`, it's
technically doing it wrong.

So, since we now can't build it that way, how do we? With `pkgsCross`!

We can now use cross-compilation to give the "correct" build environment
for the build

```
 $ nix-build -A pkgsCross.aarch64-multiplatform.mox-boot-builder
```

This is why we had to set
`HOSTCC=${buildPackages.stdenv.cc.targetPrefix}cc`; we needed to refer
to the compiler for the host platform (the one the build is happening
on).

But why? Why are we doing all this when it worked just fine?

For correctness. By marking the platform as `aarch64-linux`, as the
binary is expected to be used on, we're ensuring other parts of the
build will be able to cope with it correctly.

Co-authored-by: Samuel Dionne-Riel <samuel@dionne-riel.com>