ruby: Fix cross-compilation with yjit
Otherwise it will produce `yjit.a` with the default target for the
`rustc` compiler, which in turn turns out to be the build platform?

It could be configured using the `RUST` environment variable, and
since it's used without quoting it would expand the arguments added
to it.

Using `RUST = "rust --target ..."` I think would be  undesirable for
cross-compiling (and native), since it will save those arguments in
`lib/ruby/*/*/rbconfig.rb`. Though that might be fine too. I guess.

Note that 3.2 breaks differently. I haven't investigated it, since it's
not the default anymore, and will eventually be dropped.

Signed-off-by: Samuel Dionne-Riel <samuel@dionne-riel.com>